Task_Jay_Vaghela/TASK_1_2/LiveFurnish/requirements.txt
Django==3.1.5
django-phonenumber-field[phonenumbers]
djangorestframework
mysql